+++ Changes.md
@@ -1,6 +1,15 @@
 
 
 #
+### **+05:30 11:18:09 PM 29-08-2024, Thursday**
+
+  - Added CLI argument parser.
+  - Added `parseArguments()`.
+  - Added `gerbers` command to export Gerber files.
+  - `generateGerbers()` function now accepts the input PCB file name and the output directory path.
+  - New Version `0.0.2`.
+
+#
 ### **+05:30 10:45:47 PM 29-08-2024, Thursday**
 
   - Gerber files exported successfully.